dev/requirements.txt
fastapi
uvicorn
pillow
tensorflow
python-multipart
fastapi[standard]
